+++ requirements/backend/src/auth/auth.service.ts
@@ -9,7 +9,7 @@ import { UserDetails } from './utils/UserDetails';
 export class AuthService {
 	constructor (private userService: UserService) {}
 	async validateUser(details: UserDetails): Promise<User> {
-		let user: User = await this.userService.getUserByLogin(details.login);
+		let user: User = await this.userService.findUserByLogin(details.login);
 
 		if (user)
 			return (user);
